Subinterface selection based on Multiple Macs
As of now, in Vcenter mode, when a packet is received on physical
interface with Vlan tag, it is assigned to subinterface based on the
Source Mac of the packet and Vlan tag. This does not work incase of Vrrp
protocol, as Vrrp enabled VM's send packets with source Mac as Vrrp
group MAc. To ensure that even these packets are assigned to right
subinterface, more than one MAc is added to parent interface which
results in same subinterface.

drv_add and drv_add_sub_interface routines are manipulated to ensure
that they parent interface's db is manipulated multiple times, rather
once.

Change-Id: Id3111dddd8e931b640740ef9b471c105dc1e63cf
closes-bug: #1669518