Flow support for IPV6 fragments
Currently the IPV6 fragments are not flow processed. Flow module
identifies the packet as V6 frgments and treats it as "Flow Forward".
This results in failure of some features which are relying on Flow
procesing such as Ecmp, VRF Translation, V6 Nat and also the flow
action(If the flow action for head fragment is Drop, the subsequent
fragments should be dropped, rather being forwarded). To make the flow
processing for V6 fragments, Fragment Assembler needs to be extended to
support V6 as they can be out order.

Fragment Assembler Changes:

Same fragment table is extended even for V6 without creating
a new table. The fragment key has been modified to contain 2 uint64_t
fields to hold both V4 and V6 IP addresses. The V4 IP gets stored in
lower 64 bit filed and V6 gets stored in both 64 bits fields. Both
source and destintaion IP addresses are stored in same way. The V6
fragment identifier is 32 bit legth unlike V4 which is 16 bit field. The
fragment key identifier is also modified to have 32 bit identifier. Rest
of the whole fragment assembler processing is kept same without any
modification

Checksum changes:

Currently if the packet is received on Frabric interface with MplsoGre
encapsulation, the checksum validation is for inner packets. If the
inner packet is V6 fragment, checksum can not be validated as the TCP
checksum is for the complte V6 packet rather for that paricular
fragment. To validate the checksum in this case, all the IPV6 fragments
corresponding to that TCP segment need to be reassembled and
checksummed. To avoid the reassembly of these V6  fragments, skb's
"ip_summed" is NOT marked as "CHECKSUM_UNNECESSARY" to force the
checksum computation on VM

Pulling the headers to Head skb:

V6 fragments come with IPV6 fragment header both from VM and on Fabric
interface. So the IPV6 fragment header is pulled to head skb for
handling in all the encapsulations. If the packet is ICMPV6 error, the
payload IPV6 header can be fragment and in this case the payload
fragment header has been pulled

Reverse flow:

Agent, while writing the forward flow action, links the flow to reverse
flow using reverse flow key as it is not aware of the reverse flow
index. Vrouter does the lookup for reverse flow using the key and links
the flow by itself. But for V6, Vrouter is wrongly doing a look up on
the forward flow key rather on reverse flow to retrieve the
reverse flow. This also has been rectified as part of this commit

Change-Id: I71bd5a74da4810346aadabce888bc72179a1dac2
partial-bug: #1702675