Merge the following 4.0 commits into master
Author: Raja Sivaramakrishnan <raja@juniper.net>
Date:   Mon Sep 11 15:40:24 2017 -0700

Do not return error if qemu doesn't send a call fd

If the guest is running a DPDK application, qemu 2.9 on Redhat
doesn't send a call fd (as guest doesn't need to be interrupted).
Detect this condition and don't return error.

Author: Raja Sivaramakrishnan <raja@juniper.net>
Date:   Tue Aug 22 14:29:40 2017 -0700

Enable inner packet checksum offload only on NICs which are known to support it

Author: Raja Sivaramakrishnan <raja@juniper.net>
Date:   Mon Aug 21 15:44:45 2017 -0700

Add support for Broadcom 25G NIC in DPDK vrouter

Author: Raja Sivaramakrishnan <raja@juniper.net>
Date:   Tue Aug 22 12:18:34 2017 -0700

Handle mbuf chains in vhost0 packets when using KNI

DPDK 17.02 has support to handle mbuf chains in the KNI kernel module.
Setting the mbuf pkt_len to a lower value than the sum of the data
lengths of the mbufs in a chain caused a crash in the KNI kernel module.

Change-Id: I80779c886b195721c6d27e8c2bb064c7409ac804
Closes-bug: 1716196,1712429,1711779,1708789