Update the VIF's IP/IP6 addresses even in VIF change
Currently the IP and IP6 addresses of the VIF are copied from the
request structure only while doing add of the interface. The change of
the interface is not updating it. For the subinterfaces, Agent updates
the IP/6 addresses as part of VIF change. As the addresses are not
copied to VIF as part of VIF, the source IP of the mirrored packet is
not going with right IP.

As a fix, the VIF is updated with IP and IP6 addresses as part of VIF
change as well

Change-Id: I76fdf2590c8c2b997443f02afe5cfa6c2af28a09
closes-bug: #1717730