Update the VIF's IP/IP6 addresses even in VIF change
Currently the IP and IP6 addresses of the VIF are copied from the
request structure only while doing add of the interface. The change of
the interface is not updating it. For the subinterfaces, Agent updates
the IP/6 addresses as part of VIF change. As the addresses are not
copied to VIF as part of VIF, the source IP of the mirrored packet is
not going with right IP.

As a fix, the VIF is updated with IP and IP6 addresses as part of VIF
change as well

Change-Id: I5515e31b9a48b1e771c8c856ad7c08e29432f6e8
closes-bug: #1717730