Dont Evict a flow if HoldQ Exists
While processing the Ecmp nexthop, if no component Nexthop can be
chosen and if active flow entry exists, we create a holdq for that flow
without marking the flow action as Hold and trapping the packet to
Agent. While packets are held in this holdq, there can be an RST packet
from either of the directions. If there is an RST in reverse flow, we
start evicting the reverse flow. The process of Eviction is aborted only
if the reverse flow's action is marked as Hold. In this case as action
is not marked as Hold, eviction continues, and eventually both the flows
get evicted without holdq being cleaned up. This results in buffer and
holdq memory not being freed, though the flows are marked as "Not
Active".

As a fix, eviction is aborted not only if the flow action is Hold, but
also if the holdq exists.

Change-Id: I96d79c1bcabbe4894b80fe8fb9977dc932901dc7
closes-bug: #1719253