Subject the Vhost packets to flow processing if policy enabled on Vhost
If ACL's are configured on Vhost interface, the packets need to be
subjected to flow processing. These packets would be un-tunneled
packets. The Vhost route in VRF 0 points to L3 Receive Nh. This NH can
not be marked with policy bit as we do not want to create Flows for
outer IP fields of Tunneled packets which also get processed by L3
receive NH. Also if Policy on Vhost is disabled and if L3 receive NH is
enabled with Relaxed policy, packets have to be flow processed  for
link local processing. For this reason, policy bit on Receive NH's
outgoing interface is verified to decide whether to invoke flow processing
for packets

Change-Id: I86b8ea2ecdf1460bb0d3599675d84ad46dfcf1e8
closes-bug: #1711045