DPDK: Fix for KNI crash in pkt_drop_stats()
Make the vif argument as NULL in pkt_drop_stats() in KNI send_burst().
There may be cases when vif is deleted after the packet is sent to KNI
and when such packet hits this code, there may be a crash since the
function tries to access the deleted vif.

Change-Id: If180909a1b9828b97310329feebd32371e965c11
Closes-Bug: 1714156