Mark the BFD and healthcheck packets
BFD and Health check packets from VM and Agent need not be flow
processed. BFD packets are identified with a specific UDP destination
port and healthcheck packets are identified as ICMP with checksum of
FFFF. As FFFF can be a right checksum for an ICMP packet, only those
packets whose checksum is wrong and still marked as FFFF is considered
an healthcheck packet. To validate the checksum, a linux dependent
checksum callback is introduced to handle the case of packet being
spread across multiple buffers (skb's).

Change-Id: I626f8eeebe366d4e5ceb9400a020246ff90eab03
closes-bug: #1714059