Huge pages support in Vrouter kernel module
Vrouter currently allocates contiguous memory for Flow and Bridge
tables and this forces the module to be inserted into kernel at the
boot up time as huge memory is likely to be presetnt at the boot of the
sytem. This requirement is taken out with the huge pages support.

Provisioning:
If the huge page support is intended in the system, provisioning of the
compute node takes care of enabling the required number of huge pages in
the system. After enabling the huge pages successfully, the hugetable
file system is mounted at a specific location and creates the required
files (2 files as of now - vrouter_mem1 and vrouter_mem2). The location
of these files is made avaialble to Agent.

Agent changes:
Agent looks for the huge files at the specific path provided by
provisioning (through agent configuration file) and mmap()s the 1G size
of data. If succeeds, Agent shares this virtual memory address with
Vrouter kernel module through netlink socket. If the huge files do not
exist, or if the mmap fails, still it communicates the failure of this
to Vrouter kernel module. This messaging is done in blockig way, so that
no other configuration is added to Vrouter till this is communicated.

Vrouter changes:
1) The module_init routine is changed in such a way that, the modules like
   fib and flow which require huge contiguous amount of memory do not
   allocate any memory in their init routines.

2) A new call back "mem" is introduced which allocates the required
   memory and is invoked after succesfully initialising the hugepages
   when Agent communicates it.

3) The modules like interface etc are initialised as usual so that
   addition of fabric and Vhost interfaces are created normally and
   pushed to cross connect mode till the other configuration is
   available

4) When Agent communicates the virtual memory address, Vrouter pins
   these pages using get_user_pages(). The "mem" callback is invoked
   which allocates the memory either from huge pages or using regular
   btables

5) As the momeory allocate from the huge pages is intended for the
   entire life time of Vrouter module free() routinues are not provided
   for these memory segments. These pages are freed from the system when
   the module is removed from the kernel.

6) If agent restarts, memeory is not reinitilised again and proper
   return codes are shared with Agent

Change-Id: I942ed42463d7a0aba7bf857df7c62d9c95b83056
closes-bug: #1728925