Calculate Ecmp component NH for non-IP packets
Currently the Ecmp component NH is calculated for IP/IP6 packets only as
Ecmp has been supported for only for L3 packets. With the support of L2 Ecmp, the
component NH needs to be chosen even for L2 packets. As there is no Flow
available for L2 packets, the hash is calculated on Ether dst, src and
VRF

Change-Id: I61085c729da9633630088604c1a6b8db5897bba8
closes-bug: #1732285