Dont mirror Vlan tag in Vmware environment
Currently, if there is a sub interface created with a Vlan Tag on any
parent interface and if mirroring is enabled on subinterface, the Vlan
Tag of the subinterface is also carried in the mirrored packet. Incase
of Vmware compute, all the VMIs are configred in Vrouter as
subinterfaces with Pvlan. And if the mirroring is enabled on these
VMI's, as the VMI's are subinterfaces, even these vlans are carried in
the mirroring. But the expectation is not to carry these Vlan tags as
these Vlan tags are Contrail specific and not seen by the user.

As a fix, there is a new VIF flag introduced "VIF_FLAG_MIRROR_NOTAG"
which is set by Agent on Vxmware VMI's only. If these flags is seen on
VIF, Vrouter discards the  Vlan Tag before mirroring the packets

Change-Id: Ic46921cfd1f1e318b39d2045101f0878ee7ab79d
closes-bug: #1711459