Remove Ecmp checks in ARP processing
When L2 Ecmp is not in place, packets to/from Ecmp source are forced to
L3 packets by replying ARP requests using Vrouter's and Vhost's MAc.
This is no longer required once we have L2 Ecmp in place.

So all ARP processing which specially deals with Ecmp source side is not
reqired any more. This processing is removed from the current ARP
handling

Change-Id: I484da69b3c173891b915fb86b81c2e57d711b18a
closes-bug: #1733811