Do not calculate fragment values for mirrored packets
If fragmentes are being processed, fragment assembler maintains the
statistics of framents in the fragment entry to decide when to delete
the fragment. When mirroring is enabled, the mirrored packet gets
processes earlier than real packet and these fragment calculations are
invoked on the mirrored packet leading to fragment entry being deleted
if the mirrored packet is the last fragment of the packet. Once the
fragment entry is deleted and mirrored packet is out of system, the real
packet gets processes and this packet will not have a matching fragment
entry to do the flow processing resulting in packet being discarded.

As a fix, the fragment calculations are invoked only when the real
packets are processed rather for mirrored packets.

Change-Id: I30b3092622f8c661c6bebaf5ccbff6c9621cc3dc
closes-bug: #1739602
(cherry picked from commit c9bf93807028d387e52b373ecd1cba9ce9d48688)
(cherry picked from commit 11b441b018808999d111c3e7c02938f27f84239b)