Do not add wrong port "0" fat flow config as part of Fat flow config delete
When the deleted fat flow config contains port 0, due to wrong marking
of the port in the port fat flow config, port 0 is added as fat flow config.
This is leading to invalid fat flow config being added with proto 0 and all ports
creating wrong flows.

As a fix, the deleted ports are marked appropriately in the old configuration

Change-Id: I54ec7f59d52eb2740bd1796c02f1f82e04af2979
closes-bug: #1757366