Limiting the number of forwarding cores used by dpdk vrouter
DPDK supports and detects up 128 cores, vrouter only 64. Various buffers are sized to this 64 limit, e.g. lcores in vr_dpdk_global.
If over 64 are enabled, they can become indices in these arrays causing, overflows. Limiting the number of fwd cores that can be used so that we don't get into this situation.

Change-Id: Ie1a6eaf1eb7e8fc9d72a65d255babcd7bf119659
Closes-Bug: #1759851