DPDK: ixgbe_alloc_rx_queue_mbufs(): RX mbuf alloc failed
When a vhost0 notification (like mtu change, mac addr change etc).
is received when the interface is down, we were stopping the PMD
uncleanly resulting in errors. We don't need to stop or do anything
on the PMD. We just needed that function to set the MTU and nothing
else.

Change-Id: I965e403f6774a75f47a6c11a7a0634426722595a
Closes-Bug: 1761369