Choose more specific fat flow config over wild card
On a vif, per a given protocol, there can be fat flow configuration
for an explicit port and as well for port "0". In this case, the fat
flow config for more specific port should be considered first while
making the flow entry rather port 0.

As a fix, the lookup for specific is carried first before making for
port 0 to apply the config

Change-Id: I82d2097ee71a62d3768b3ae5a4b29bd207b11c4a
closes-bug: #1761486
(cherry picked from commit 07d9719984483e7a29468bd351c54d0036022bcd)