This is merge from - https://review.opencontrail.org/#/c/41260/
Changes are introduced to handle the ipsec decrypted packets over the
VTI (Virtual Tunnel Interface), an L3 interface. IPSec kernel will
process the encrypted packet, decrypt it and then write it to the
VTI interface.

vrouter has the VTI interface plumbed to it and process the RX of
the VTI interface. This interface will receive the decrypted L3VPN
packet. The L3 interface has the L3 packet. Ethernet header is read
from skb.

Closes-Bug: 1760242

Change-Id: I60e4a20e5b436094d33ddc207d60c6d5275fe466