Appky the rewrite data for Multicast packets from Vhost interface
With the support of VM's on Fabric VN, the multicast support on Fabric
and Vhost is implicit. But for the packets coming from Vhost interface,
the ethernet header rewrite data is never applied as the dst mac
constructed by HostOS is considered always correct. When the packets is
destined to ECMP destination on Fabric VN, destination MAC of the packet
from Vhost would have been having Vrouter's MAC (as Vrouter would have
replied with self for Ecmp). This ethernet header needs to be changed as
per the component nexthop chosen from Ecmp.

As a fix, if the packet is from Vhost interface, and being driven sent
out for Ecmp destination, rewrite data is applied

Change-Id: I02bd367b734dec39297e9166f6606fe68af023f8
closes-bug: #1764939