Send ARP resonse on the physical interface without a bridge lookup
In case of VM's in Fabric VN the ARP requests from Baremetal servers can
reach Vrouter on fabric interface for VM. These arp replies should not
be bridged as bridge tables in VRF 0 are not populated. ARP replies need
to transmitted as is on the receiving interface.

As a fix, the ingress interface itself is used to send ARP responses on
physical interface

Change-Id: Ibcd0eaaf78b1483bed1f8f95ea85dc08a7b791ae
closes-bug: #1764969