DPDK GSO: Fix for performance issue seen in TIM
The condition whether to segment a packet or not was wrong. The correct
condition is that if TSO seg size that the VM is giving in virtio
header is lesser than the packet size, it needs to be segmented.

Change-Id: I97cf8a23407dac08030c3df575b5a3522b06fd2e
Closes-Bug: 1772648
(cherry picked from commit 43d884991513479f689e931b86486f26dae06f80)