DPDK: Performance issue in TSO
There was a side effect in the earlier commit that Rx direction will
have an issue due to the will_segment condition. In Rx, the packet gets
GROed and hits the same code path and has to evaluate will_segment as
false since the GROed packet needs to goto the VM. But it was evaluating
to true. Hence added an extra check.

Change-Id: I6216aef7cad3b372d1aa8286e4dc5c7f723dec2e
Closes-Bug: 1772648