btable: optimize entry address calculation
The btable uses division and modulo calculations, which are quiet
expensive in the most platforms, for the entry address calculations.

In case the allocation limit of a btable is power of 2 number,
the calculations can be optimized to use simple bit calculations.

Optimize the entry address calculations in the aforementioned case.

Partial-Bug: #1781402
Change-Id: Idfc56e14a6d499d23a306866cfa463f95eee1151
Signed-off-by: Matan Azrad <matan@mellanox.com>