[WIP] Ensure that all headers are in single allocation on Windows
Note: this code is not yet tested, but it should be in reviewable state.

When receiving a fragmented packets from a container,
the protocol header may be in separate allocation than
ethernet + IP headers. dp-core assumes that all these
headers are in single chunk of memory so sometimes
we need to reallocate the buffer.

Change-Id: I9372cf79dbe2cf66d47e9e42daec134933d96c94
Partial-Bug: #1773203