vr_htable_find_hentry to bypass vr_btable_get
In htable_find_hentry when iterate the entries in a bucket,
vr_btable_get is called. However, when the btable_partition
is 1, we can just simply shift the pointer to avoid the
repeated btable calculation. This is especially useful
for non-kernel path since the btable parition is always 1

This commit checks parition count and if it is 1 then do
pointer shift instead of repeatedly calling vr_btable_get.

Change-Id: If1d860f2c6ac07182ffad812670490955cfb371c
Partial-Bug: #1769570
Signed-off-by: Yipeng Wang <yipeng1.wang@intel.com>