windows: Introduce WinTxPostprocess in win_if_tx
This change introduces a first iteration of WinTxPostprocess, which would be
unit testable in user space. Right now it implements a logic from lines 261-266
of windows/vr_host_interface.c. Although it's not used right now, it is planned
that WinTxPostprocess will replace most of win_if_tx code in subsequent changes.

To make first iteration of WinTxPostprocess compilable in user space,
the following functions had to be copied and modified:

- fix_ip_csum_at_offset,
- zero_ip_csum_at_offset,
- fix_csum,
- fix_tunneled_csum,
- fix_ip_v4_csum.

Change-Id: I2a4ac48d7c422063194ba26265534f81a456b1a0
Partial-Bug: #1796663