Update MTU of vif when it connects to the vswitch
This commit is related in data race between vRouter extension and connection
of virtual interface to vswitch.

After container is created, hyper-V exposes assigned virtual interface
for the container, but the vif connection to the vswitch is done asynchronously.

MTU for the virtual interface is updated only when it's connected to the vswitch.

There were cases when vRouter extension read MTU from vif before it was
connected, so extension had incorrect value.

The fragmentation didn't happen and the read MTU was higher then
physical interface's so the fragment was dropped.

Change-Id: I4aff05543507318fe77d21f4a76a3f02b955dad4
Closes-Bug: #1803541