dp-core: Add dynamic offload module registration
Extend the paravirtualized per-host type offloads with infrastructure
that allows arbitrary modules to plug into the framework. This closely
resembles the offload register/unregister callbacks that were present in
versions R3.1 and R4.1, with the addition of dpdk support (under Linux).

RCU is used to ensure integrity of the offload structure upon
plugging/unplugging. The registration interface exposes symbols when
building the vrouter kernel module. A framework for user space plugging
is not yet available, but dynamic loading could be used to achieve this.

Change-Id: Ifa2ea8af5b0abfeb4e732a4f68d48cf42bd8739f
Partial-Bug: #1767107
Signed-off-by: Frik Botha <frederick.botha@netronome.com>