Do not decrement TTL while doing L3 forward to agent pkt0
In BGPaaS in pkt mode, the pkt will be L3 routed with the NH
pointing to pkt0 or GW interface at one stage (before NAT).
In single hop BGPaaS case, since the TTL is decremented in
vr_forward() during L3 routing, the pkt is dropped later
since the TTL becomes 0. This leads to BGP session not
coming up at all.

Change-Id: I68d482a903de9067464d75fe269d74f9d2beaf67
Closes-Bug:#1737963
closes-jira-bug:#JCB-199874