Do not decrement TTL while doing L3 forward to agent pkt0
In BGPaaS in pkt mode, the pkt will be L3 routed with the NH
pointing to pkt0 or GW interface at one stage (before NAT).
In single hop BGPaaS case, since the TTL is decremented in
vr_forward() during L3 routing, the pkt is dropped later
since the TTL becomes 0. This leads to BGP session not
coming up at all.

Change-Id: I959dcf4e1b316a76698f2b5d954bbaf0f197b64e
closes-jira-bug:#JCB-199874