For local ECMP NHs, pick the vif from the child member NH while creating fat flows
Today, in case of local ECMP NHs, the flow gets created using the ECMP NH
as Key NH because of which the vif interface couldn't be retrieved to apply
fat flow config.
With this fix, we will check for local ECMP NH, and if so we will pick the
vif from the first member NH. If the first member vif interface has
fat flow configured, then it will be applied.

Change-Id: Icf7b029678eea9da8d0dd5416c4b22248332a8cc
partial-jira-bug: JCB-219075