Sending pckts on bare-metal doesn't break vrouter
Function ReallocateHeaders performs NBL copy for all packets - no
matter if headers are in continuous memory segment or not.
It means that if NIC driver is about to reuse original NBL in the
future, it may safely assume that it is not modified.

This change only affects Windows specific ode.

Change-Id: I5528c230db094857c7f489c7cd8a070a770fc626
Closes-JIRA-Bug: WIND-178