vRouter: Fix for SCP not working
Fix the MTU calculation. In platform dependent code, the default MTU is
taken from the linux netdevice which is 1500 bytes. But for VIF, the MTU
is 1514 bytes (max size including ether hdr). As a result, the MSS
calculation would cause the size to go beyond 1514byes and the NIC was
dropping large size ethernet packets. Fixed by adjusting the MTU to take
care of ethernet header size.

Change-Id: I377446dd6430a62b35c0d80c2bd9554b72bf3ba4
Closes-Jira-Bug: CEM-4137