NTTC issue: CEM-4275 - One of the TCP dead flow doesn't get evicted
Root cause:
===========
This is due a race condition between setting of the
Evict Candidate flag and the Eviction of the flow.

We have 2 threads here say Thread1 and Thread2.

Thread1 is in the flow defer_callback function,
while Thread2 is in the flow mark evict function.
Now consider the following sequence (Time T0 to T4)
which leads to the non eviction of the flow 100.
T0 and T3 are executed in Thread1, while T1, T2 and T4
are executed as part of Thread 2.

Lets assume the flow index for the flows as 100 and 200.

Thread 1                               Thread 2
-------------                          -------------
Defer_cb_Func()                        Mark_Evict_Func()

Time T0 -  CheckEvictFlow(100)  –
           No op, as Evict candidate
           flag is not set for flow 100

                                      Time T1 – Set Evict Candidate
                                                for flow(100)
                                      Time T2 – Set Evict Candidate
                                                for flow(200)

Time T3 – CheckEvictFlow(200) –
          Now flow 200 would get
          evicted since Evict
          Candidate flag is set.

                                      Time T4 – Schedule Defer_cb for
                                                flow(200)

Now, since flow 200 already got evicted at Time T3,
the callback would never be scheduled at time T4.
And hence, the flow 100 would never get evicted.
Was able to repro the problem by simulating the above order
using instrumented code and scapy script.

The problem is with Defer_cb_func() which should do the eviction
only when the Evict Candidate flag is set for both flows.

Fix:
====
The fix is to make sure the eviction is done only after both
the flows evict candidate flags are set.

Testing:
========
- Done UT for the code changes
- QA has qualified the fix

Change-Id: I07730ab190646260d08de6ca4fdf9bc1caf16d6e
closes-jira-bug: CEM-4275