Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > Where/how are we tracking the addition of UT case for this scenario?
> 
> I guess we need some kind of whitebox UT infra for testing these kind of scenarios

This is a race condition and it is not be possible to predictably recreate the issue. We need to instrument to code to add delays to simulate the race condition. So this bug is not a good candidate for UT.

Patch-set: 1
Reviewer: Gerrit User 232 <232@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Label: Verified=0