Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (4 comments)

Another comment. The design is still reliant on maintaining state in the kernel module as static variables. So this still makes the design prone to failure when the CLI doesn't complete a request and requests same or another core again. Since the static variable is never reset until the request for a core is complete. 
There is no good way to resolve this, unless the design does away the need to maintain state in kernel module.

Patch-set: 5
Reviewer: Gerrit User 575 <575@c72cd27d-1743-4a55-84a4-cd13b8400c52>