CEM-11421 - Orange ZRB Missing prefixes series of occurrences
Issue:

Routes were missing for couple of prefixes in the vrouter routing table.
Specifically 10.60.7.0/25 and 10.60.7.128/25 prefixes.

Root cause:

The routing table in vrouter is implemented as a mtrie data structure.
The root cause of the issue was incorrect mtrie bucket deletion logic in mtrie code.
The issue was reproduced with the following sequence of route adds and delete.

addr = 10.60.7.0 plen = 25 vrf = 4 operation = ADD/CHANGE nh_idx = 38 label = 410760
addr = 10.60.0.0 plen = 23 vrf = 4 operation = ADD/CHANGE nh_idx = 38 label = 410760
addr = 10.60.7.128 plen = 25 vrf = 4 operation = ADD/CHANGE nh_idx = 38 label = 410760
addr = 10.60.0.0 plen = 20 vrf = 4 operation = ADD/CHANGE nh_idx = 63 label = 609700
addr = 10.60.0.0 plen = 20 vrf = 4 operation = DELETE nh_idx = 51

After executing this sequence, it was observed that the 10.60.7.0/25 and
10.60.7.128/25 routes were getting deleted as part of 10.60.0.0/20 delete operation.
This was because as part of delete operation, we were deleting the mtrie bucket
containing 10.60.7.0/25 and 10.60.7.128/25 routes wrongly due to incorrect
logic/assumption in code.

Fix:

The mtrie bucket deletion logic is changed to delete the bkt using one of the
entries in the bucket instead of deleting it using the values from the route delete
request. This fixes the issue of bucket with non matching prefix length being deleted.

Verification:

The fix was verified by rerunning the same sequence again and
checking if the 10.60.7.0/25 and 10.60.7.128/25 routes are present or not.
A UT script has been added for the same.
The fix also passed full vrouter regression in 3.2.3.x branch.

Change-Id: If558d05bb666b48bb2ae48a34ec03692c57f74fb