Socket leak when VM stop/start In loop
When VIF is added but not connected, vru_cl->vruc_fd is not added to the
fd list even though it is created. This can happen when VM is stopped.
In this case, vr_uvhost_del_fds_by_arg() would not close
the fd. This leads to socket leak, so added fcntl() call in
vr_uvhost_del_client() to check if vruc_fd is closed or not.
Closes-jira-bug: CEM-9285

Change-Id: Ia55e44b8830b96e8313691dc77873bfd133e1469