Vrouter design to handle flows
Issue:

Customer has a VM connected to a vrouter and connected to a F5 loadbalancer through fabric
interface of vrouter.
In the customer scenario, there was an out of order TCP RST packet received from load balancer,
for which vrouter was closing the flow without checking the TCP seqnum. But, the end client
checks for it properly and ignores the TCP RST packet and sends a challenge ack as per RFC.
However since vrouter has already closed the flow wrongly, subsequent packets for the flow
are dropped. This breaks the TCP protocol between the client and server. The server thinks
the connection is closed but the client still has the connection open.

Solution:

There are 2 options to fix this:-

Option 1:
- Implement the TCP seqnum check as per RFC and drop the RST packet in vrouter itself.

Option 2:
- Do not close the flow on receiving the TCP RST packet and let the flow age out on its own
  to avoid interfering with TCP state machine (This is earlier contrail behaviour before TCP flow
  eviction was implemented)

Option 2 was chosen because with Option 1, there is still a possibility that we may close the flow
on valid RST packet, but later the packet may get dropped down the line. And after closing the flow
at vrouter, we can again land into the same situation (i.e client server mismatched states).
This is because vrouter will not allow the TCP state machine to complete fully.

To maintain backward compatibility, a flag vr_close_flow_on_tcp_rst is added to allow configuring
the old behaviour (in case someone still wants it)

NOTE: The default behaviour without the knob configured is to not close the TCP flow on RST.

Closes-Jira-Bug: CEM-10847
Change-Id: I700a416525c0ae79da4d7fc98ce7cf76a8fcc326