CEM-14660 - Avoid using /dev/vr_bridge and /dev/flow devices in agent             when kernel is using hugepages
On newer ubuntu kernels, a traceback related to bad PTE entries is seen
when agent is exiting. The same is also seen when flow and rt commands are used.
The traceback is seen when unmap for /dev/vr_bridge and /dev/flow is being done.
It is not clear what is causing this. The same is not seen on older kernels used
by RHEL and Centos.

To avoid this issue, agent/flow/rt need not use /dev/vr_bridge and flow devices
since in kernel with hugepages mapped, agent/flow/rt can directly use the mmaped
/dev/hugepages/bridge and /dev/hugepages/flow files.

Vrouter changes:
1. Get the hugepage file path sent by agent and store it in the internal hugepage
   cfg struct along with other hugepage information.
2. Given the stored hugepage file path to flow and rt command so that they can mmap
   the tables properly.

Testing:
1. Tested with kernel mode without hugepages enabled
2. Tested with kernel (RHEL/Centos and Ubuntu) with hugepages enabled
3. Tested with DPDK compute

partial-jira-bug: CEM-14660
Change-Id: I2a916c23088221d8a2ee236bac3a4bde78384142