CEM-13716: Remote ECMP NH doesn't work With Fabricforwarding
Root cause:
In case of local ECMP processing, as in fip is on one VM on local
compute and another VM on remote compute, when the pkt is chosen
to be sent to remote compute (by agent selecting the remote ecmp idx),
the fabric vif rewrite (L2 rewrite) was not happening due to
some special checks in the code.

Fix:
Changed the check condition to also check for dst mac of the
pkt. If dmac of the pkt is equal to vhost0 mac, then the pkt
is being routed through vrouter and we need to honor the NH
found during route lookup. In such cases, we need to do the
L2 rewrite using NH information.

closes-jira-bug: CEM-13716
Change-Id: I90dc16b5e48aba6156eddfef7b1486b6faf10741