Support not moving phys interface attributes to vhost0 Closes-Jira-Bug: CEM-15411
The motivation behind the change is to remove the maintainence
associated with moving interface attributes from the physical
interface associated with the vhost0 interface to the vhost0
interface across distros or distro versions. Furthermore,
Ignition used by CoreOS considers this config as immutable.

The patch changes the RX code path to send to the kernel stack
from the input interface context after passing the packet thru.
vrouter. On the TX side, the physical interface 'ndo_start_xmit'
handler, associated with vhost0 is overridden. This allows
packets that the kernel sends out the physical interface directly
to be intercepted and sent thru. vrouter, before being potentially
sent out the physical interface.

There are some known additional changes to be done in
agent-functions.sh, init_vhost0() to not move the physical
interface attributes (eg) IP address) to vhost0.

Also, potentially vtest changes, which is TBD, at this time.

Signed-off-by: Darrell Ball <dball@juniper.net>
Depends-On: I6724088fc441f88b248c18d64fc42a6646ece31e
Change-Id: Ibbdfc71e5d01ddf5661901056f4ebfbef51a57d4