Support not moving phys interface attributes to vhost0 Closes-Jira-Bug: CEM-15411
Vrouter side code to support not needing to move physical
interface attributes, notably IP address, from the physical
interface to the vhost0 interface. The rational is to avoid
support issues associated with different interface managers
in various Linux distros and/or versions of those distros.
For example, Ignition in CoreOS considers the interface
attributes as immutable after being set.

On RX, the packets that entering the physical interface are
passed thru. the vrouter pipeline and if are destined to be
sent to the Linux stack via vhost0, are sent from the physical
interface context to the Linux stack.

On TX, the ndo_start_xmit() function of the physical interface
is overridden, which for packets being sent directly from the
kernel, will pass packet thru. the vrouter pipeline first and
then use the original ndo_start_xmit to send packets thru. the
physical interface.

Dependent script changes, located in another repo, are noted here.

There may be some vtest changes as well, which is TBD.

Signed-off-by: Darrell Ball <dball@juniper.net>
Depends-On: I6724088fc441f88b248c18d64fc42a6646ece31e
Change-Id: Ibbdfc71e5d01ddf5661901056f4ebfbef51a57d4