Health check fails to work after VM reboot
Issue:
When the VM is rebooted, server stops running and health check fails momentarily, due to which the route to VM points to nh_discard. When the first TCP SYN packet is sent, it is trapped by agent while flow is programmed. Before the packet is enqueued NH of the packet is set to NULL. After flushing the packet, vr_inet_route_lookup() is done on packet with NAT IP whose route is already pointed to nh_discard. Due to this the first SYN packet is being dropped. By the time retransmission of this packet happens, it reaches health check timeout of 1 sec and connection closes. This keeps on repeating

Fix:
After flushing the packet, before the IP NAT happens, we do a vr_inet_route_lookup() which gives us the NH associated with non-NAT IP which prevents the first SYN packet drop.

Closes-Jira-Bug: CEM-11226
Change-Id: I9193242918d42bcf1bc7b6590884a3da9f785d20