DPDK Fragmentation resulting in mbuf leak and stops traffic
Issue:
When fragmented packets are sent between VMs, mbufs which are allocated from Indirect mempool are
not being freed resulting in packet drops after the mempool is exhausted. When a mempool is
created, a specified number of cache mbufs are created and initialised which are later allocated asindirect mbufs during fragmentation. On each block in the list, rte_pktmbuf_init() function is
called to initialise mbuf with default values. In this function mbuf_size is calculated using
mbuf_priv_size which is uninitialised and has some random large value which results in large
mbuf_size. And later when memset is called to reinitialise the memory equivalent to mbuf_size, all the contents and references of further mbufs are wiped, due to which they are not initialised.
These uninitialised mbufs are not freed after allocation as it checks for values like mempool and
ref_cnt which are not set in these mbufs.

	Fragmentation is working fine on R3.2 which is using DPDK 17.02. Though it has the same
unintialised random mbuf_priv_size value used in calculating mbuf_size, memset() is called on
elt_size(128 in indirect mempool) as opposed to mbuf_size(large value) in DPDK 18.05.

Fix:
We have to pass rte_pktmbuf_pool_init() as constructor during mempool creation. In this function
mbuf_priv_size is set to zero which prevents mbuf_size being large during mbuf initialisation.

closes-jira-bug: CEM-18101
Change-Id: I1881077261876645146fb3c47ea15004a87e878f