CEM-17211 Overlay ICMP Type03/Code04(for PMTUD) is not forwarded
Double commit to R1912

Root cause:

The flow sport and dport was taken as 0 for the ICMP error pkt,
while Agent used the sport as the inner ICMP Echo req Id field.
Because of this there was mismatch in the flow creation between
agent and vrouter leading to the issue.

The fix is to use the inner ICMP Echo req id as sport in vrouter as well.

closes-jira-bug: CEM-17211
Change-Id: I05c18b4ab1e3812645fd7eb4baa7a9515535d17b