Problem statement:  Customer observed BGPaaS stuck in OPEN / RST loop on Ericsson VNF.
Desription: After establishing a BGPaas session with the the controller in session 1 , a hold timer gets expired due to unknown reason and VNF closes the BGP connection ( Even though TCP FIN was sent , but the connection is in the process of Closure). At this time , VNF again establishes the connection with the controller in session 2, vrouter creates the flows for session 2 and updates its flow table. At this time TCP retransmission takes place for the FIN packet , as the TCP state machine at the client have not received the ACK for the FIN of session 1 . Thus triggering the paired flow creation  for session 1 and causes the deletion of forward flow for session2.

Having the corrupted flow stack , when controller sends BGP packets to VNF , Vrouter NATs it to wrong port , upon detecting the packets with wrong port VNF issues RST to reset the connection . This happens in loop.

Root Cause : The reason behind this issue is Creation/Addition of flows when FIN packet is received at the vrouter.

Fix/Solution: Not to create the flows whenever FIN is received at the vrouter.

Closes-jira-bug: CEM-20284

Signed-off-by: NagarajM R <mrn@nodem1.englab.juniper.net>
Change-Id: I792da8b7cd1891d4bbbb5e52e8f2b32c3854c617